Add files via upload
This is a solar system simulation where the planets are distanced upto scale and one can zoom in and out to see the desired planets scale. The simulation can be paused by pressing the P key on your keyboard and one can note down the exact position where the planets are from the sun. 
Accesing it is very simple, just run the code and pygame will run a window on your screen, game mechanics are extremely simple and take only three commands, pausing, zooming in-out and quiting. 
It can be made more complex in the future by showing the trails of orbits or having a starry background.